Move QuoteStyle out of the account object
