+++ legacy/account/src/main/java/app/k9mail/legacy/account/Account.kt
@@ -600,11 +600,6 @@ open class Account(
         return uuid.hashCode()
     }
 
-    enum class QuoteStyle {
-        PREFIX,
-        HEADER,
-    }
-
     companion object {
         /**
          * Fixed name of outbox - not actually displayed.
