+++ legacy/core/src/main/java/com/fsck/k9/message/MessageBuilder.java
@@ -15,7 +15,7 @@ import com.fsck.k9.mail.internet.AddressHeaderBuilder;
 import com.fsck.k9.mail.internet.Headers;
 import timber.log.Timber;
 
-import app.k9mail.legacy.account.Account.QuoteStyle;
+import app.k9mail.legacy.account.QuoteStyle;
 import app.k9mail.legacy.account.Identity;
 import com.fsck.k9.K9;
 import app.k9mail.legacy.message.controller.MessageReference;
