+++ legacy/core/src/main/java/com/fsck/k9/message/quote/TextQuoteCreator.kt
@@ -1,6 +1,6 @@
 package com.fsck.k9.message.quote
 
-import app.k9mail.legacy.account.Account.QuoteStyle
+import app.k9mail.legacy.account.QuoteStyle
 import com.fsck.k9.CoreResourceProvider
 import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.Message
