+++ legacy/core/src/test/java/com/fsck/k9/message/quote/TextQuoteCreatorTest.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.message.quote
 
 import app.k9mail.core.android.testing.RobolectricTest
-import app.k9mail.legacy.account.Account.QuoteStyle
+import app.k9mail.legacy.account.QuoteStyle
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import com.fsck.k9.TestCoreResourceProvider
