+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/settings/account/AccountSettingsFragment.kt
@@ -21,6 +21,7 @@ import app.k9mail.feature.launcher.FeatureLauncherActivity
 import app.k9mail.feature.launcher.FeatureLauncherTarget
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.NO_OPENPGP_KEY
+import app.k9mail.legacy.account.QuoteStyle
 import com.fsck.k9.account.BackgroundAccountRemover
 import com.fsck.k9.activity.ManageIdentities
 import com.fsck.k9.activity.setup.AccountSetupComposition
@@ -196,8 +197,8 @@ class AccountSettingsFragment : PreferenceFragmentCompat(), ConfirmationDialogFr
     private fun initializeQuoteStyle() {
         findPreference<Preference>(PREFERENCE_QUOTE_STYLE)?.apply {
             setOnPreferenceChangeListener { _, newValue ->
-                val quoteStyle = Account.QuoteStyle.valueOf(newValue.toString())
-                notifyDependencyChange(quoteStyle == Account.QuoteStyle.HEADER)
+                val quoteStyle = QuoteStyle.valueOf(newValue.toString())
+                notifyDependencyChange(quoteStyle == QuoteStyle.HEADER)
                 true
             }
         }
